move xbox live progress menu to proper folder
- rename module to proper name
