+++ xlive/Blam/Engine/interface/screens/screen_xbox_live_task_progress_dialog.h
