+++ xlive/H2MOD/GUI/imgui_integration/MOTD.cpp
@@ -2,13 +2,13 @@
 
 #include "imgui_handler.h"
 
-#include "H2MOD.h"
+#include "Blam/Engine/interface/screens/screen_xbox_live_task_progress_dialog.h"
+
 #include "H2MOD/Modules/Input/ControllerInput.h"
 #include "H2MOD/Modules/Input/KeyboardInput.h"
 #include "H2MOD/Modules/Input/Mouseinput.h"
 #include "H2MOD/Modules/Input/PlayerControl.h"
 #include "H2MOD/Modules/Shell/Startup/Startup.h"
-#include "H2MOD/Modules/UI/XboxLiveTaskProgress.h"
 #include "Util/Hooks/Hook.h"
 #include "XLive/xnet/IpManagement/XnIp.h"
