+++ xlive/H2MOD/Modules/CustomMenu/Accounts/c_account_create_menu.cpp
@@ -3,11 +3,11 @@
 #include "c_account_create_menu.h"
 
 #include "Blam/Engine/memory/data.h"
+#include "Blam/Engine/interface/screens/screen_xbox_live_task_progress_dialog.h"
 
 #include "H2MOD/Modules/Accounts/Accounts.h"
 #include "H2MOD/Modules/Accounts/AccountLogin.h"
 #include "H2MOD/Modules/Accounts/AccountCreate.h"
-#include "H2MOD/Modules/UI/XboxLiveTaskProgress.h"
 #include "H2MOD/Modules/OnScreenDebug/OnscreenDebug.h"
 
 #include "../c_error_menu.h"
